방 생성 시 제목과 비밀번호 설정 추가
플레이어 점수에 따른 랭킹 시스템 버그 수정

+++ Source/BumperCart/PlayerState/MainPlayerState.h
@@ -68,7 +68,7 @@ private:
     void OnRep_DroppedItemCount();
 
     UPROPERTY(ReplicatedUsing = OnRep_PlayerScore, VisibleAnywhere,  Category = "PlayerStats")
-    int32 PlayerScore = 0;
+    float PlayerScore = 0.f;
 
     UPROPERTY(ReplicatedUsing = OnRep_Rank, VisibleAnywhere,  Category = "PlayerStats")
     int32 Rank = 0;
